tests: move data providers into DataProviders directory
Relates to #28
